code climat && review changes

+++ src/main/java/ru/mail/polis/service/physhock/ServiceImpl.java
@@ -1,11 +1,16 @@
 package ru.mail.polis.service.physhock;
 
 import com.google.common.base.Charsets;
-import com.google.common.collect.Iterables;
-import one.nio.http.*;
+import one.nio.http.HttpServer;
+import one.nio.http.HttpServerConfig;
+import one.nio.http.HttpSession;
+import one.nio.http.Param;
+import one.nio.http.Path;
+import one.nio.http.Request;
+import one.nio.http.RequestMethod;
+import one.nio.http.Response;
 import one.nio.net.Socket;
 import one.nio.server.AcceptorConfig;
-import one.nio.server.RejectedSessionException;
 import org.jetbrains.annotations.NotNull;
 import ru.mail.polis.Record;
 import ru.mail.polis.dao.DAO;
@@ -20,11 +25,24 @@ import java.nio.charset.Charset;
 import java.util.Iterator;
 import java.util.concurrent.Executor;
 
+/**
+ * Implementation of Service.
+ */
 public class ServiceImpl extends HttpServer implements Service {
 
+    private static final Response INTERNAL_ERROR = new Response(Response.INTERNAL_ERROR, Response.EMPTY);
+    private static final Response BAD_REQUEST = new Response(Response.BAD_REQUEST, Response.EMPTY);
     private final DAO dao;
     private final Executor executor;
 
+    /**
+     * Server constructor.
+     *
+     * @param port     server port
+     * @param dao      dao ( RocksDB dao)
+     * @param executor executor service
+     * @throws IOException super thrower
+     */
     public ServiceImpl(final int port, final DAO dao, final Executor executor) throws IOException {
         super(getConfig(port), dao);
         this.dao = dao;
@@ -49,37 +67,31 @@ public class ServiceImpl extends HttpServer implements Service {
     @Path("/v0/status")
     @RequestMethod(Request.METHOD_GET)
     public Response getStatus(@NotNull final Request request) {
-        return Iterables.isEmpty(request.getParameters())
-                ? new Response(Response.OK, "I am alive!\n".getBytes(Charset.defaultCharset()))
-                : new Response(Response.BAD_REQUEST, Response.EMPTY);
+        return new Response(Response.OK, "I am alive!\n".getBytes(Charset.defaultCharset()));
     }
 
     /**
      * Method returns data if exists.
      *
      * @param id key value
-     * @return data
      */
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_GET)
     public void getData(@Param(value = "id", required = true) final String id, final HttpSession session) {
-        executor.execute(() -> {
-            try {
-                geniusCheck(id);
-                final ByteBuffer result = dao.get(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
-                final Response response = new Response(Response.OK, ByteBufferUtils.getByteArray(result));
-                putResponseToSession(session, response);
-            } catch (NoSuchElementExceptionLite e) {
-                final Response response = new Response(Response.NOT_FOUND, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IOException e) {
-                final Response response = new Response(Response.INTERNAL_ERROR, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IllegalArgumentException e) {
-                final Response response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-                putResponseToSession(session, response);
-            }
-        });
+        if(geniusCheck(id, session)) {
+            executor.execute(() -> {
+                try {
+                    final ByteBuffer result = dao.get(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
+                    final Response response = new Response(Response.OK, ByteBufferUtils.getByteArray(result));
+                    putResponseToSession(session, response);
+                } catch (NoSuchElementExceptionLite e) {
+                    final Response response = new Response(Response.NOT_FOUND, Response.EMPTY);
+                    putResponseToSession(session, response);
+                } catch (IOException e) {
+                    putResponseToSession(session, INTERNAL_ERROR);
+                }
+            });
+        }
     }
 
     /**
@@ -87,62 +99,60 @@ public class ServiceImpl extends HttpServer implements Service {
      *
      * @param request data
      * @param id      id
-     * @return status of operation
      */
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_PUT)
     public void putData(final Request request, @Param("id") final String id, final HttpSession session) {
-        executor.execute(() -> {
-            try {
-                geniusCheck(id);
-                dao.upsert(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)), ByteBuffer.wrap(request.getBody()));
-                final Response response = new Response(Response.CREATED, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IOException e) {
-                final Response response = new Response(Response.INTERNAL_ERROR, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IllegalArgumentException e) {
-                final Response response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-                putResponseToSession(session, response);
-            }
-        });
+        if(geniusCheck(id, session)) {
+            executor.execute(() -> {
+                try {
+                    dao.upsert(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)), ByteBuffer.wrap(request.getBody()));
+                    final Response response = new Response(Response.CREATED, Response.EMPTY);
+                    putResponseToSession(session, response);
+                } catch (IOException e) {
+                    putResponseToSession(session, INTERNAL_ERROR);
+                }
+            });
+        }
     }
 
     /**
      * Method deletes data with defined id.
      *
      * @param id id
-     * @return status of operation
      */
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_DELETE)
     public void deleteData(@Param("id") final String id, final HttpSession session) {
-        executor.execute(() -> {
-            try {
-                geniusCheck(id);
-                dao.remove(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
-                final Response response = new Response(Response.ACCEPTED, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IOException e) {
-                final Response response = new Response(Response.INTERNAL_ERROR, Response.EMPTY);
-                putResponseToSession(session, response);
-            } catch (IllegalArgumentException e) {
-                final Response response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-                putResponseToSession(session, response);
-            }
-        });
+        if(geniusCheck(id, session)) {
+            executor.execute(() -> {
+                try {
+                    dao.remove(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
+                    final Response response = new Response(Response.ACCEPTED, Response.EMPTY);
+                    putResponseToSession(session, response);
+                } catch (IOException e) {
+                    putResponseToSession(session, INTERNAL_ERROR);
+                }
+            });
+        }
     }
 
+    /**
+     * Method get data in requested range.
+     *
+     * @param start   start form
+     * @param end     end by
+     * @param session http session which wil
+     */
     @Path("/v0/entities")
     @RequestMethod(Request.METHOD_GET)
     public void getRange(@Param(value = "start", required = true) final String start,
                          @Param(value = "end") final String end,
-                         final HttpSession session, final Request request) {
-        executor.execute(() -> {
-            try {
-                geniusCheck(start);
+                         final HttpSession session) {
+        if (geniusCheck(start, session)) {
+            executor.execute(() -> {
                 final ByteBuffer from = ByteBuffer.wrap(start.getBytes(Charsets.UTF_8));
-                ByteBuffer to = end == null || end.isEmpty()
+                final ByteBuffer to = end == null || end.isEmpty()
                         ? null
                         : ByteBuffer.wrap(end.getBytes(Charsets.UTF_8));
                 try {
@@ -152,27 +162,26 @@ public class ServiceImpl extends HttpServer implements Service {
                 } catch (IOException e) {
                     throw new UncheckedIOException("Session troubles", e);
                 }
-            } catch (IllegalArgumentException e) {
-                final Response response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-                putResponseToSession(session, response);
-            }
-        });
+            });
+        }
     }
 
     @Override
-    public void handleDefault(final Request request, final HttpSession session) throws IOException {
-        session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
+    public void handleDefault(final Request request, @NotNull final HttpSession session) throws IOException {
+        session.sendResponse(BAD_REQUEST);
     }
 
     @Override
-    public HttpSession createSession(Socket socket) throws RejectedSessionException {
+    public HttpSession createSession(final Socket socket) {
         return new ChunkedSession(socket, this);
     }
 
-    private void geniusCheck(final String id) throws IllegalArgumentException {
-        if ("".equals(id)) {
-            throw new IllegalArgumentException("String ? return; // nothing to do");
+    private boolean geniusCheck(final String id, final HttpSession session) {
+        final boolean equals = "".equals(id);
+        if (equals) {
+            putResponseToSession(session, BAD_REQUEST);
         }
+        return !equals;
     }
 
     private void putResponseToSession(@NotNull final HttpSession session, final Response response) {
